various small fixes

+++ src/model/post_db.php
@@ -49,7 +49,7 @@ class PostDB {
     $sql = "
     SELECT post.id, post.user_id, user.username, post.type, post.title, post.description, post.location, post.created_datetime FROM post
     JOIN user ON user.id = post.user_id
-    WHERE visible = 1
+    WHERE visible = 1 AND post.id = $id
     ORDER BY post.id DESC
     ";
     $post = $db->fetch($sql);
@@ -58,7 +58,7 @@ class PostDB {
     $sql = "
     SELECT comment.id, username, user_id, created_datetime, content FROM comment
     JOIN user ON user.id = comment.user_id
-    WHERE comment.object_id = 1 AND comment.object_type = 'post'
+    WHERE comment.object_id = $id AND comment.object_type = 'post'
     ORDER BY comment.created_datetime DESC
     ";
     $comments = $db->fetch_multiple($sql);
@@ -67,14 +67,14 @@ class PostDB {
     $sql = "
     SELECT * FROM user_like
     JOIN user ON user.id = user_like.user_id
-    WHERE user_like.object_id = 1 AND user_like.object_type = 'post'
+    WHERE user_like.object_id = $id AND user_like.object_type = 'post'
     ";
     $likes = $db->fetch_multiple($sql);
 
     // get images
     $sql = "
     SELECT file_name FROM image
-    WHERE object_id = 1 AND object_type = 'post'
+    WHERE object_id = $id AND object_type = 'post'
     ";
     $images = $db->fetch_multiple($sql);
 
@@ -82,7 +82,7 @@ class PostDB {
     $user_id = (isset($_SESSION['user_id'])) ? $_SESSION['user_id'] : -1;
     $sql = "
     SELECT COUNT(*) AS liked FROM user_like
-    WHERE user_id = $user_id AND object_id = 1 AND object_type = 'post'
+    WHERE user_id = $user_id AND object_id = $id AND object_type = 'post'
     ";
     $liked = $db->fetch($sql)['liked'];
 
@@ -131,8 +131,20 @@ class PostDB {
     return $result;
   }
 
-  public static function delete($post_id) {
-        $db = SimpleDB::Singleton();
+  public static function delete($user_id, $post_id) {
+    $db = SimpleDB::Singleton();
+
+    // check if user is owner of post
+    $sql = "
+    SELECT COUNT(*) AS count FROM post
+    WHERE id = $post_id AND user_id = $user_id
+    ";
+    $result = $db->fetch($sql);
+
+    if ($result['count'] == 0) {
+      return json_encode(['state' => 'error', 'message' => 'You are not the owner of this post.']);
+    }
+
     $sql = "DELETE FROM post WHERE id = $post_id";
     $result = $db->query($sql);
 
@@ -142,6 +154,10 @@ class PostDB {
 
     // delete all likes
     $sql = "DELETE FROM user_like WHERE object_id = $post_id AND object_type = 'post'";
+    $result = $db->query($sql);
+
+    // delete all images
+    $sql = "DELETE FROM image WHERE object_id = $post_id AND object_type = 'post'";
     $result = $db->query($sql);
 
     return $result;
